+++ crates/criticalup-cli/src/commands/install.rs
@@ -5,9 +5,11 @@ use std::path::{Path, PathBuf};
 
 use criticaltrust::integrity::IntegrityVerifier;
 use criticaltrust::manifests::{Release, ReleaseArtifactFormat};
+use criticalup_core::download_server_cache::DownloadServerCache;
 use criticalup_core::download_server_client::DownloadServerClient;
 use criticalup_core::project_manifest::{ProjectManifest, ProjectManifestProduct};
 use criticalup_core::state::State;
+use tokio::fs::read;
 
 use crate::errors::Error;
 use crate::errors::Error::{IntegrityErrorsWhileInstallation, PackageDependenciesNotSupported};
@@ -18,12 +20,19 @@ pub const DEFAULT_RELEASE_ARTIFACT_FORMAT: ReleaseArtifactFormat = ReleaseArtifa
 pub(crate) async fn run(
     ctx: &Context,
     reinstall: bool,
+    offline: bool,
     project: Option<PathBuf>,
 ) -> Result<(), Error> {
     // TODO: If `std::io::stdout().is_terminal() == true``, provide a nice, fancy progress bar using indicatif.
     //       Retain existing behavior to support non-TTY usage.
 
     let state = State::load(&ctx.config).await?;
+    let maybe_client = if !offline {
+        Some(DownloadServerClient::new(&ctx.config, &state))
+    } else {
+        None
+    };
+    let cache = DownloadServerCache::new(&ctx.config.paths.cache_dir, &maybe_client).await?;
 
     // Get manifest location if arg `project` is None
     let manifest_path = ProjectManifest::discover_canonical_path(project.as_deref()).await?;
@@ -37,7 +46,7 @@ pub(crate) async fn run(
         let abs_installation_dir_path = installation_dir.join(product.installation_id());
 
         if !abs_installation_dir_path.exists() {
-            install_product_afresh(ctx, &state, &manifest_path, product).await?;
+            install_product_afresh(ctx, &state, &cache, &manifest_path, product).await?;
         } else {
             // Check if the state file has no mention of this installation.
             let does_this_installation_exist_in_state = state
@@ -46,7 +55,7 @@ pub(crate) async fn run(
             if !does_this_installation_exist_in_state || reinstall {
                 // If the installation directory exists, but the State has no installation of that
                 // InstallationId, then re-run the install command and go through installation.
-                install_product_afresh(ctx, &state, &manifest_path, product).await?;
+                install_product_afresh(ctx, &state, &cache, &manifest_path, product).await?;
             } else {
                 // If the installation directory exists AND there is an existing installation with
                 // that InstallationId, then merely update the installation in the State file to
@@ -67,9 +76,16 @@ pub(crate) async fn run(
     Ok(())
 }
 
+#[tracing::instrument(level = "debug", skip_all, fields(
+    manifest_path = %manifest_path.display(),
+    installation_id = %product.installation_id(),
+    release = %product.release(),
+    product = %product.name(),
+))]
 async fn install_product_afresh(
     ctx: &Context,
     state: &State,
+    cache: &DownloadServerCache<'_>,
     manifest_path: &Path,
     product: &ProjectManifestProduct,
 ) -> Result<(), Error> {
@@ -77,17 +93,15 @@ async fn install_product_afresh(
     let release = product.release();
     let installation_dir = &ctx.config.paths.installation_dir;
     let abs_installation_dir_path = installation_dir.join(product.installation_id());
-    let client = DownloadServerClient::new(&ctx.config, state);
-    let keys = client.get_keys().await?;
+    let keys = cache.keys().await?;
 
-    // TODO: Add tracing to support log levels, structured logging.
     tracing::info!("Installing product '{product_name}' ({release})",);
 
     let mut integrity_verifier = IntegrityVerifier::new(&keys);
 
     // Get the release manifest for the product from the server and verify it.
-    let release_manifest_from_server = client
-        .get_product_release_manifest(product_name, product.release())
+    let release_manifest_from_server = cache
+        .product_release_manifest(product_name, product.release())
         .await?;
     let verified_release_manifest = release_manifest_from_server.signed.into_verified(&keys)?;
 
@@ -103,10 +117,8 @@ async fn install_product_afresh(
         .await?;
 
     for package in product.packages() {
-        tracing::info!("Downloading component '{package}' for '{product_name}' ({release})",);
-
-        let response_file = client
-            .download_package(
+        let package_path = cache
+            .package(
                 product_name,
                 release_name,
                 package,
@@ -114,18 +126,10 @@ async fn install_product_afresh(
             )
             .await?;
 
-        // Archive file path, path with the archive extension.
-        let package_name_with_extension =
-            format!("{}.{}", package, DEFAULT_RELEASE_ARTIFACT_FORMAT);
-        let abs_artifact_compressed_file_path: PathBuf =
-            abs_installation_dir_path.join(&package_name_with_extension);
-
-        // Save the downloaded package archive on disk.
-        tokio::fs::write(&abs_artifact_compressed_file_path, response_file.clone()).await?;
-
         tracing::info!("Installing component '{package}' for '{product_name}' ({release})",);
+        let package_data = read(package_path).await?;
 
-        let decoder = xz2::read::XzDecoder::new(response_file.as_slice());
+        let decoder = xz2::read::XzDecoder::new(package_data.as_slice());
         let mut archive = tar::Archive::new(decoder);
         archive.set_preserve_permissions(true);
         archive.set_preserve_mtime(true);
@@ -147,8 +151,6 @@ async fn install_product_afresh(
                 );
             }
         }
-
-        clean_archive_download(&abs_artifact_compressed_file_path).await?;
     }
 
     let verified_packages = integrity_verifier
@@ -170,11 +172,6 @@ fn check_for_package_dependencies(verified_release_manifest: &Release) -> Result
             return Err(PackageDependenciesNotSupported(package.package.clone()));
         }
     }
-    Ok(())
-}
-
-async fn clean_archive_download(abs_artifact_compressed_file_path: &PathBuf) -> Result<(), Error> {
-    tokio::fs::remove_file(abs_artifact_compressed_file_path).await?;
     Ok(())
 }