+++ crates/criticalup-core/src/lib.rs
@@ -3,10 +3,10 @@
 
 pub mod binary_proxies;
 pub mod config;
+pub mod download_server_cache;
 pub mod download_server_client;
 pub mod errors;
 pub mod project_manifest;
-
 pub mod state;
 
 mod utils;